# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for Timeful.",
        "title": "Timeful API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3002/api",
    "paths": {
...
        "/analytics/user/{email}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Gets the user by email",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User email",
                        "name": "email",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        },
...
        "/events": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Creates a new event",
                "parameters": [
                    {
                        "description": "Timed events require the complete canonical slot contract; day-only events require dates",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "activeSlots": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                "attendees": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                "blindAvailabilityEnabled": {
                                    "type": "boolean"
                                },
                                "dates": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                "daysOnly": {
                                    "type": "boolean"
                                },
                                "enabledSlots": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                "eventTimezone": {
                                    "type": "string"
                                },
                                "isSignUpForm": {
                                    "type": "boolean"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "notificationsEnabled": {
                                    "type": "boolean"
                                },
                                "remindees": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                "sendEmailAfterXResponses": {
                                    "type": "integer"
                                },
                                "signUpBlocks": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/models.SignUpBlock"
                                    }
                                },
                                "slotGeneration": {
                                    "$ref": "#/definitions/models.SlotGeneration"
                                },
                                "timedRecurrence": {
                                    "$ref": "#/definitions/models.TimedRecurrence"
                                },
                                "type": {
                                    "$ref": "#/definitions/models.EventType"
                                },
                                "when2meetHref": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "eventId": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/user/add-apple-calendar-account": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Adds an apple calendar account",
                "parameters": [
                    {
                        "description": "Object containing the email and app password of the apple account",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/user/add-google-calendar-account": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Adds a new calendar account",
                "parameters": [
                    {
                        "description": "Object containing the Google authorization code and scope",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "string"
                                },
                                "scope": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/user/add-ics-calendar-account": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Adds an ICS calendar account",
                "parameters": [
                    {
                        "description": "Object containing the feed URL and label of the ICS calendar",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "feedUrl": {
                                    "type": "string"
                                },
                                "label": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/user/add-outlook-calendar-account": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Adds a new outlook calendar account",
                "parameters": [
                    {
                        "description": "Object containing the Outlook authorization code and scope",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "string"
                                },
                                "scope": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.User": {
            "type": "object",
            "properties": {
                "_id": {
                    "description": "Profile info",
                    "type": "string"
                },
                "calendarAccounts": {
                    "description": "CalendarAccounts is a mapping from {`email_CALENDARTYPE` =\u003e CalendarAccount} that contains all the\nadditional accounts the user wants to see google calendar events for",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/models.CalendarAccount"
                    }
                },
                "calendarOptions": {
                    "description": "Calendar options",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.CalendarOptions"
                        }
                    ]
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "hasCustomName": {
                    "description": "Whether the user has set a custom name for themselves, i.e. don't change their name when they sign in",
                    "type": "boolean"
                },
                "isPremium": {
                    "type": "boolean"
                },
                "lastName": {
                    "type": "string"
                },
                "numEventsCreated": {
                    "type": "integer"
                },
                "picture": {
                    "type": "string"
                },
                "primaryAccountKey": {
                    "description": "The calendarAccountKey of the account the user first signed in with",
                    "type": "string"
                },
                "stripeCustomerId": {
                    "description": "Stripe customer ID",
                    "type": "string"
                },
                "timezoneOffset": {
                    "type": "integer"
                }
            }
        },
...
    }
}